[10.x] Add `HandlesPotentiallyTranslatedString` trait  (#47488)
* Add HandlesPotentiallyTranslatedString trait

* formatting

* rename file

---------
